Refactor to facilitate additional transports
