+++ lib/model_context_protocol/server.rb
@@ -20,7 +20,7 @@ module ModelContextProtocol
     def start
       configuration.validate!
       logdev = configuration.logging_enabled? ? $stderr : File::NULL
-      StdioTransport.new(logger: Logger.new(logdev), router:).begin
+      Transports::Stdio.new(logger: Logger.new(logdev), router:).begin
     end
 
     private
